+++ izmk-core/src/main/kotlin/ovo/xsvf/izmk/gui/widget/AbstractWidget.kt
@@ -12,15 +12,15 @@ abstract class AbstractWidget(val screen: GuiScreen) {
     val mc by lazy { IZMK.mc }
     val logger: Logger = LogManager.getLogger(javaClass)
 
-    abstract fun getHeight(): Int
-    abstract fun draw(screenWidth: Int, screenHeight: Int,
+    abstract fun getHeight(): Float
+    abstract fun draw(screenWidth: Float, screenHeight: Float,
                       renderX: Float, renderY: Float,
                       fontMulti: FontMultiDraw, rectMulti: PosColor2DMultiDraw,
                       partialTicks: Float)
-    open fun mouseClicked(mouseX: Double, mouseY: Double, isLeftClick: Boolean) {}
+    open fun mouseClicked(mouseX: Float, mouseY: Float, isLeftClick: Boolean) {}
     open fun keyPressed(keyCode: Int, scanCode: Int): Boolean { return false }
 
-    fun drawDefaultBackground(rectMulti: PosColor2DMultiDraw, renderX: Float, renderY: Float, screenWidth: Int) {
+    fun drawDefaultBackground(rectMulti: PosColor2DMultiDraw, renderX: Float, renderY: Float, screenWidth: Float) {
         rectMulti.addRectGradientHorizontal(
             renderX, renderY,
             screenWidth - 2 * 5f, getHeight().toFloat(),
